Fixed RTL for Xubuntu slideshow (LP: #560220).
This required that bullet points be drawn with the usual list-style-image CSS property, instead of the :before pseudo-class.

Updated Xubuntu's bullet-point.png so the RTL version is identical to the LTR version (except flipped). Added 1px padding to top of image, since we are no longer able to fine-tune its alignment within CSS.

Visuals should be 1-to-1 identical, unless you're using an RTL language, in which case they're much improved :)